+++ NEWS.md
@@ -8,6 +8,7 @@
 * Refactored `tm_t_rsp` with new function `template_rsp`.
 * Refactored `tm_t_tte` with new function `template_tte`.
 * Refactored `tm_t_summary` with new function `template_summary`.
+* New module `tm_t_ancova` with new function `template_ancova`.
 * New utility function: `styled_expr`.
 * New utility function `pipe_expr` to concatenate expressions.
 * New utility function `add_expr` to add expressions to a list of expressions.